chore: simplify dist generation
only produce commonjs and module
