+++ packages/@lwc/node-reactions/package.json
@@ -3,13 +3,13 @@
     "version": "1.0.2",
     "description": "Node Reactions utility to detect synchronous mutations of the DOM",
     "license": "MIT",
-    "main": "dist/commonjs/index.js",
-    "module": "dist/modules/es2017/node-reactions.js",
+    "main": "dist/node-reactions.cjs.js",
+    "module": "dist/node-reactions.js",
     "typings": "dist/types/index.d.ts",
     "scripts": {
         "clean": "rm -rf dist/",
         "test": "jest",
-        "build": "tsc && rollup --config ./scripts/rollup/rollup.config.js"
+        "build": "tsc  --emitDeclarationOnly && rollup --config ./scripts/rollup/rollup.config.js"
     },
     "files": [
         "dist/"
