+++ packages/@lwc/node-reactions/scripts/rollup/rollup.config.js
@@ -4,8 +4,31 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-module.exports = [
-    ...require('./rollup.config.es-and-cjs'),
-    ...require('./rollup.config.umd.dev'),
-    ...require('./rollup.config.umd.prod'),
-];
+const path = require('path');
+const typescript = require('rollup-plugin-typescript');
+
+const { version } = require('../../package.json');
+const entry = path.resolve(__dirname, '../../src/index.ts');
+const targetDirectory = path.resolve(__dirname, '../../dist');
+const banner = `/**\n * Copyright (C) 2018 salesforce.com, inc.\n */`;
+const footer = `/** version: ${version} */`;
+
+function generateTargetName({ format }) {
+    return ['node-reactions', format === 'cjs' ? '.cjs' : '', '.js'].join('');
+}
+
+function rollupConfig({ format }) {
+    return {
+        input: entry,
+        output: {
+            file: path.join(targetDirectory, generateTargetName({ format })),
+            name: 'NodeReactions',
+            format,
+            banner,
+            footer,
+        },
+        plugins: [typescript({ target: 'es2017', typescript: require('typescript') })],
+    };
+}
+
+module.exports = [rollupConfig({ format: 'es' }), rollupConfig({ format: 'cjs' })];
