--- packages/@lwc/node-reactions/scripts/rollup/util.js
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: MIT
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
- */
-const COMPAT_SUFFIX = '_compat';
-const DEBUG_SUFFIX = '_debug';
-const PROD_SUFFIX = '.min';
-
-function generateTargetName({ prod, proddebug }) {
-    return ['node-reactions', proddebug ? DEBUG_SUFFIX : '', prod ? '.min' : '', '.js'].join('');
-}
-
-module.exports = {
-    COMPAT_SUFFIX,
-    DEBUG_SUFFIX,
-    PROD_SUFFIX,
-    generateTargetName,
-};
